Added the same changes for the offer page
