+++ apps/portal/src/components/pages/SignupPage.js
@@ -9,6 +9,7 @@ import InputForm from '../common/InputForm';
 import {ValidateInputForm} from '../../utils/form';
 import {getSiteProducts, getSitePrices, hasOnlyFreePlan, isInviteOnlySite, freeHasBenefitsOrDescription, hasOnlyFreeProduct, getFreeProductBenefits, getFreeTierDescription, hasMultipleNewsletters, hasFreeTrialTier, isSignupAllowed} from '../../utils/helpers';
 import {ReactComponent as InvitationIcon} from '../../images/icons/invitation.svg';
+import {interceptAnchorClicks} from '../../utils/links';
 
 export const SignupPageStyles = `
 .gh-portal-back-sitetitle {
@@ -529,17 +530,6 @@ class SignupPage extends React.Component {
         const errorClassName = this.state.errors?.checkbox ? 'gh-portal-error' : '';
 
         const className = `gh-portal-signup-terms ${errorClassName}`;
-
-        const interceptAnchorClicks = (e) => {
-            if (e.currentTarget.contains(e.target)) {
-                const anchor = e.target.closest('a');
-
-                if (anchor) {
-                    e.preventDefault();
-                    window.open(anchor.href, '_blank');
-                }
-            }
-        };
 
         return (
             <div className={className} onClick={interceptAnchorClicks}>
